quest pre requisite helper method
